My Parts

+++ app/src/main/java/com/example/project_g05/ParkFragment.kt
@@ -153,7 +153,7 @@ class ParkFragment : Fragment(), OnMapReadyCallback {
 
                         // Add markers on map for each national park
                         for (park in parkList) {
-                            val latLng = LatLng(park.latitude.toDouble(), park.longitude.toDouble())
+                            val latLng = LatLng(park.latitude, park.longitude)
                             mMap.addMarker(
                                 MarkerOptions()
                                     .position(latLng)
